extract functions to proper places
