+++ dbcon/execplan/rewrites.h
@@ -1,0 +1,30 @@
+/* Copyright (C) 2022 MariaDB Corporation
+
+   This program is free software; you can redistribute it and/or
+   modify it under the terms of the GNU General Public License
+   as published by the Free Software Foundation; version 2 of
+   the License.
+
+   This program is distributed in the hope that it will be useful,
+   but WITHOUT ANY WARRANTY; without even the implied warranty of
+   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+   GNU General Public License for more details.
+
+   You should have received a copy of the GNU General Public License
+   along with this program; if not, write to the Free Software
+   Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
+   MA 02110-1301, USA. */
+
+#pragma once
+
+#include <parsetree.h>
+#include <treenode.h>
+
+
+namespace execplan {
+
+//Walk the tree and find out common conjuctions
+execplan::ParseTree* extractCommonLeafConjunctionsToRoot(execplan::ParseTree* tree);
+
+
+} // namespace execplan
